+++ src/WeBoard.Client/WeBoard.Core/Updates/Interfaces/IUpdate.cs
@@ -4,7 +4,7 @@ namespace WeBoard.Core.Updates.Interfaces
 {
     public interface IUpdate
     {
-        public long TargetId { get;}
+        public int TargetId { get;}
         public DateTime Date { get;}
         public void Apply(ITrackable trackable);
         IUpdate GetCancelUpdate();
